requirements.txt
# Core
pyyaml>=6.0

# PDF Processing
PyMuPDF>=1.24.0
Pillow>=10.0.0

# LLM Providers
anthropic>=0.49.0
openai>=1.55.0
httpx>=0.28.0

# GUI
customtkinter>=5.2.2

# Utils
aiofiles>=24.0.0
